+++ CHANGELOG.md
@@ -14,8 +14,11 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 - kokkos@3.7.02
 
 ### Added
+- Added support for unstructured topologies with mixed elements types (for example, hexs and tets).
+- Added support for `pyramid` and `wedge` elements.
 - Added a `topologies` option to the relay extract. This allows you to select which topologies are saved. This option can be used with the existing `fields` option, the result is the union of the selected topologies and fields.
 - Added `near_plane` and `far_plane` to the camera details provided in Ascent::info()
+
 
 ## [0.9.3] - Released 2024-05-11
 ### Preferred dependency versions for ascent@0.9.3